+++ component/discovery/kuma/kuma_test.go
@@ -1,0 +1,36 @@
+package kuma
+
+import (
+	"testing"
+
+	"github.com/grafana/river"
+	"github.com/stretchr/testify/require"
+)
+
+func TestRiverConfig(t *testing.T) {
+	var exampleRiverConfig = `
+	server = "http://kuma-control-plane.kuma-system.svc:5676"
+
+	refresh_interval = "10s"
+	fetch_timeout    = "50s"
+`
+
+	var args Arguments
+	err := river.Unmarshal([]byte(exampleRiverConfig), &args)
+	require.NoError(t, err)
+}
+
+func TestBadRiverConfig(t *testing.T) {
+	var exampleRiverConfig = `
+	server = "http://kuma-control-plane.kuma-system.svc:5676"
+	tls_config {
+		ca_file = "/path/to/ca_file"
+		ca_pem = "not a real pem"
+	}
+`
+
+	// Make sure the TLSConfig Validate function is being utilized correctly
+	var args Arguments
+	err := river.Unmarshal([]byte(exampleRiverConfig), &args)
+	require.ErrorContains(t, err, "at most one of ca_pem and ca_file must be configured")
+}